requirements.txt
alembic==1.13.3 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
fastapi==0.115.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
grpcio-tools==1.66.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.2 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.6 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.5 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.1 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.12" and python_version < "4.0"
psycopg2==2.9.9 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic[dotenv]==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
pydub==0.25.1 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
setuptools==75.1.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.35 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.38.6 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.31.0 ; python_version >= "3.12" and python_version < "4.0"
yandex-speechkit==1.5.0 ; python_version >= "3.12" and python_version < "4.0"